refactor: pass reflex/tactician to xor function instead of hard coding them
Determining if a hand has only one card out of a given set of cards is a common use case between discarding and retaining. This PR renames the XOR function to "Count" (because that's what it's actually doing), and creates a new parameter: a list of the cards to consider. This way, we can pass different cards for discarding and retaining, without re-writing the function.