feat: ignore cards with Retain when retaining
It turns out I hadn't thought through the logic completely. We don't want to prioritize retaining cards with Retain. We actually want to ignore them when considering which card to retain, because they retain themselves. So in the end, we don't need to reuse those methods. This PR adds the "ignore cards with Retain" logic to the auto-retainer, and sets those two methods back to private.